% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the HSIcal Shiny App}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}}
}
\value{
No return value, called for side effects (launches Shiny app)
}
\description{
Launches the hyperspectral image calibration tool.
}
\examples{
if (interactive()) {
  run_app()
}
}
